Rename: chatSocket 위치 변경 #165
