+++ frontend/react-app/src/pages/directMessage/container.tsx
@@ -1,6 +1,6 @@
 import { useEffect, useState } from "react";
 import DirectMessagePageView from "./view";
-import { chatSocket } from "@router/socket/chatSocket";
+import { chatSocket } from "@src/utils/sockets/chatSocket";
 import { ChatType, EnterDmReturnType, MessageType } from "@src/types";
 import { useParams } from "react-router-dom";
 import { useRecoilState, useRecoilValue } from "recoil";
